refactor: extract gui widgets (WorkerThread, ModernButton, NoFocusDelegate, CoreTempDialog) into gui_widgets
